factorized my code and put it in the events directory. added dashToDOM function for welcom page
